Initial commit of the tissue encoder device, C# application and output files that were last used.
This device and C# application code is of an older revision as the later stuff was lost in the transition of dev PCs. This is more of a starting point if changes need to be made rather than the code that created the latest output files.

Generally speaking, the files in the OutputFiles folder is what should be the latest and should be used for testing.